feat: add publishing cache
Adds functionality to cache messages, that are send while broker downtime (while the client is recovering) to publish them after the connection is reestablished.
The cache itself is an interface, that can be implemented by the user. The user could for example implement caching to somting like redis or some other store of choice.